feat(ui): spin the running state and quiet the spawn marker
- The below-editor strips (subagents and workflows) now show the
  braille spinner for running work instead of a static ●, in step with
  the dashboard and takeover headers. The frames and cadence move to
  shared/spinner.ts so every surface animates identically.
- The spawn result card's bright success ● drops to dim: a spawn is a
  beginning, not a success, and the strip's spinner carries the running
  state from there.
