+++ extensions/subagents/navigation.ts
@@ -9,6 +9,7 @@ import {
   unreadActivityCounts,
   type ActivityCounts,
 } from "../shared/activity-status.ts";
+import { spinnerFrame } from "../shared/spinner.ts";
 import { sanitizeTerminalText } from "../shared/terminal-text.ts";
 import { formatElapsed, type SubagentSnapshot } from "./src/domain.ts";
 import { contextPercent } from "./src/format.ts";
@@ -59,9 +60,13 @@ function statusColor(status: SubagentSnapshot["status"]) {
   return "error" as const;
 }
 
-/** One status glyph per run state; doubles as the focus marker when selected. */
-function statusGlyph(snapshot: SubagentSnapshot, theme: Theme) {
-  if (snapshot.status === "running") return theme.fg("warning", "●");
+/**
+ * One status indicator per run state; doubles as the focus marker when
+ * selected. Running spins, in step with the dashboard and takeover headers.
+ */
+function statusGlyph(snapshot: SubagentSnapshot, theme: Theme, now: number) {
+  if (snapshot.status === "running")
+    return theme.fg("warning", spinnerFrame(now));
   if (snapshot.status === "done") return theme.fg("success", "✓");
   return theme.fg("error", "x");
 }
@@ -100,7 +105,7 @@ export class SubagentStripWidget {
     const { snapshot, counts } = entry;
     const glyph = this.strip.focused
       ? this.theme.fg("accent", "❯")
-      : statusGlyph(snapshot, this.theme);
+      : statusGlyph(snapshot, this.theme, Date.now());
     const titleText = normalizeSubagentTitle(snapshot.title, snapshot.id);
     const title = this.strip.focused
       ? this.theme.bold(this.theme.fg("accent", titleText))
